Extract registration views into their own module.
This is a pure refactor.
