+++ project/thscoreboard/users/urls.py
@@ -6,6 +6,7 @@ from users import other_views
 from users.views import accept_invite
 from users.views import batch_invite_csv
 from users.views import profile
+from users.views import register
 from users.views import request_invite
 
 from thscoreboard import settings
@@ -13,10 +14,10 @@ from thscoreboard import settings
 app_name = 'users'
 
 urlpatterns = [
-    path('register', other_views.register, name='register'),
-    path('registration_success', other_views.registration_success, name='registration_success'),
-    path('preregistered', other_views.preregistered, name='preregistered'),
-    path('verify_email/<str:token>', other_views.verify_email, name='verify_email'),
+    path('register', register.register, name='register'),
+    path('registration_success', register.registration_success, name='registration_success'),
+    path('preregistered', register.preregistered, name='preregistered'),
+    path('verify_email/<str:token>', register.verify_email, name='verify_email'),
     path('accept_invite/<str:token>', accept_invite.accept_invite, name='accept_invite'),
     path('profile', profile.profile, name='profile'),
     path('batch_invite', batch_invite_csv.batch_invite, name='batch_invite'),
